feat: removed dupped active paths
